Some audio stuff that appears to be workng

+++ libempr/audioplayback.h
@@ -10,7 +10,7 @@
 #define CW_DEMO_DELTA 0x10
 #define CW_DEMO false
 
-void playback_init(uint8_t *buf, size_t len, uint32_t sample_period);
+void playback_init(uint16_t *buf, size_t len, uint32_t sample_period);
 void playback_play();
 void playback_pause();
 void playback_deinit();
